+++ nengo/builder/processes.py
@@ -138,7 +138,7 @@ class SimProcess(Operator):
 
 
 @Builder.register(Process)
-def build_process(model, process, sig_in=None, sig_out=None, mode="set"):
+def build_process(model, process, sig_in=None, sig_out=None, mode="set", seed_or_rng=0):
     """Builds a `.Process` object into a model.
 
     Parameters
@@ -153,6 +153,8 @@ def build_process(model, process, sig_in=None, sig_out=None, mode="set"):
         The output signal, or None if no output signal.
     mode : "set" or "inc" or "update", optional
         The ``mode`` of the built `.SimProcess`.
+    seed_or_rng : int or `numpy.random.RandomState`
+        The parent seed or random number generator to use if the seed is not set.
 
     Notes
     -----
@@ -171,7 +173,10 @@ def build_process(model, process, sig_in=None, sig_out=None, mode="set"):
         if sig_in is not None
         else rc.float_dtype
     )
-    state_init = process.make_state(shape_in, shape_out, model.dt, dtype=dtype)
+    state_rng = process.get_rng(seed_or_rng, offset=1)  # offset matches `Process.apply`
+    state_init = process.make_state(
+        shape_in, shape_out, model.dt, rng=state_rng, dtype=dtype
+    )
     state = {}
     for name, value in state_init.items():
         state[name] = Signal(value)